+++ workspace/solutions/kotlin/lib/src/test/kotlin/s0/s3/s0/s4/s_0304_test.kt
@@ -1,0 +1,16 @@
+package s0.s3.s0.s4
+
+import utils.T
+import kotlin.test.Test
+import s0.s3.s0.s4.s_0304_00.NumMatrix as s_0304_00
+
+class s_0304_test {
+    @Test
+    fun test() {
+        val t = T(
+            this::class,
+            s_0304_00::class,
+        )
+        t.run()
+    }
+}