+++ workspace/solutions/rust/src/s0/s3/s0/s3/mod.rs
@@ -1,5 +1,4 @@
 pub mod s_0303_00;
-pub mod s_0303_01;
 pub mod s_0303_test;
 
 pub trait ISolution {