+++ workspace/solutions/rust/src/s0/s3/s0/s4/s_0304_test.rs
@@ -1,8 +1,8 @@
 #[cfg(test)]
 mod tests {
-    use super::super::s_0303_00;
+    use super::super::s_0304_00;
     use crate::{
-        s0::s3::s0::s3::ISolution,
+        s0::s3::s0::s4::ISolution,
         utils::{json::ValueExtension, test::t_},
     };
 
@@ -10,11 +10,11 @@ mod tests {
     fn test() -> () {
         t_(
             module_path!(), //
-            &vec![0, 1],
+            &vec![0],
             |solution, init| -> Box<dyn ISolution> {
-                let nums = init[0].to_i32_array().unwrap();
+                let matrix = init[0].to_i32_matrix().unwrap();
                 match solution {
-                    0 => Box::new(s_0303_00::NumArray::init(nums)),
+                    0 => Box::new(s_0304_00::NumMatrix::init(matrix)),
                     _ => unreachable!(),
                 }
             },
@@ -22,6 +22,8 @@ mod tests {
                 serde_json::Value::from(solution.solve(
                     input[0].to_i32().unwrap(), //
                     input[1].to_i32().unwrap(),
+                    input[2].to_i32().unwrap(),
+                    input[3].to_i32().unwrap(),
                 ))
             },
         );